copy my adapters
